requirements.txt
Flask==1.1.2
Flask-Login==0.5.0
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.4
GitPython==3.1.11
Jinja2==2.11.2
pathlib==1.0.1
setuptools==51.0.0
SQLAlchemy==1.3.20
twine==3.2.0
wheel==0.36.2
